+++ src/components/organisms/Form/FormButton/FormButton.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import Button, { IButtonProps } from '../../../atoms/Button/Button';
-import { useFormContext } from '../useForm';
+import { useFormContext } from '../hooks';
 
 export interface IFromButtonProps extends IButtonProps {
   type?: 'button' | 'reset' | 'submit';
